+++ sdmx-dl-provider-ri/src/main/java/internal/sdmxdl/ri/web/drivers/Sdmx21Driver2.java
@@ -157,6 +157,16 @@ public final class Sdmx21Driver2 implements SdmxWebDriver {
                     .build())
             .source(SdmxWebSource
                     .builder()
+                    .name("SPC")
+                    .description("Pacific Data Hub")
+                    .driver(RI_SDMX_21)
+                    .endpointOf("https://stats-nsi-stable.pacificdata.org/rest")
+                    .propertyOf(DETAIL_SUPPORTED_PROPERTY, true)
+                    .websiteOf("https://stats.pacificdata.org/?locale=en")
+                    .monitorOf("UptimeRobot", "m788488402-48b2f77f10060fcc6bb4baaa")
+                    .build())
+            .source(SdmxWebSource
+                    .builder()
                     .name("UNDATA")
                     .description("Data access system to UN databases")
                     .driver(RI_SDMX_21)